refactor: Introduce WebhdfsCore for improved service structure
