+++ core/src/services/webhdfs/writer.rs
@@ -15,11 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
+use std::sync::Arc;
+
 use bytes::Buf;
 use http::StatusCode;
 use uuid::Uuid;
 
-use super::backend::WebhdfsBackend;
+use super::core::WebhdfsCore;
 use super::error::parse_error;
 use crate::raw::*;
 use crate::services::webhdfs::message::FileStatusWrapper;
@@ -29,26 +31,26 @@ pub type WebhdfsWriters =
     TwoWays<oio::BlockWriter<WebhdfsWriter>, oio::AppendWriter<WebhdfsWriter>>;
 
 pub struct WebhdfsWriter {
-    backend: WebhdfsBackend,
+    core: Arc<WebhdfsCore>,
 
     op: OpWrite,
     path: String,
 }
 
 impl WebhdfsWriter {
-    pub fn new(backend: WebhdfsBackend, op: OpWrite, path: String) -> Self {
-        WebhdfsWriter { backend, op, path }
+    pub fn new(core: Arc<WebhdfsCore>, op: OpWrite, path: String) -> Self {
+        WebhdfsWriter { core, op, path }
     }
 }
 
 impl oio::BlockWrite for WebhdfsWriter {
     async fn write_once(&self, size: u64, body: Buffer) -> Result<Metadata> {
         let req = self
-            .backend
+            .core
             .webhdfs_create_object_request(&self.path, Some(size), &self.op, body)
             .await?;
 
-        let resp = self.backend.client.send(req).await?;
+        let resp = self.core.client.send(req).await?;
 
         let status = resp.status();
         match status {
@@ -58,14 +60,14 @@ impl oio::BlockWrite for WebhdfsWriter {
     }
 
     async fn write_block(&self, block_id: Uuid, size: u64, body: Buffer) -> Result<()> {
-        let Some(ref atomic_write_dir) = self.backend.atomic_write_dir else {
+        let Some(ref atomic_write_dir) = self.core.atomic_write_dir else {
             return Err(Error::new(
                 ErrorKind::Unsupported,
                 "write multi is not supported when atomic is not set",
             ));
         };
         let req = self
-            .backend
+            .core
             .webhdfs_create_object_request(
                 &format!("{}{}", atomic_write_dir, block_id),
                 Some(size),
@@ -74,7 +76,7 @@ impl oio::BlockWrite for WebhdfsWriter {
             )
             .await?;
 
-        let resp = self.backend.client.send(req).await?;
+        let resp = self.core.client.send(req).await?;
 
         let status = resp.status();
         match status {
@@ -84,7 +86,7 @@ impl oio::BlockWrite for WebhdfsWriter {
     }
 
     async fn complete_block(&self, block_ids: Vec<Uuid>) -> Result<Metadata> {
-        let Some(ref atomic_write_dir) = self.backend.atomic_write_dir else {
+        let Some(ref atomic_write_dir) = self.core.atomic_write_dir else {
             return Err(Error::new(
                 ErrorKind::Unsupported,
                 "write multi is not supported when atomic is not set",
@@ -97,11 +99,9 @@ impl oio::BlockWrite for WebhdfsWriter {
                 .map(|s| format!("{}{}", atomic_write_dir, s))
                 .collect();
             // concat blocks
-            let req = self
-                .backend
-                .webhdfs_concat_request(&first_block_id, sources)?;
+            let req = self.core.webhdfs_concat_request(&first_block_id, sources)?;
 
-            let resp = self.backend.client.send(req).await?;
+            let resp = self.core.client.send(req).await?;
 
             let status = resp.status();
 
@@ -110,7 +110,7 @@ impl oio::BlockWrite for WebhdfsWriter {
             }
         }
         // delete the path file
-        let resp = self.backend.webhdfs_delete(&self.path).await?;
+        let resp = self.core.webhdfs_delete(&self.path).await?;
         let status = resp.status();
         if status != StatusCode::OK {
             return Err(parse_error(resp));
@@ -118,7 +118,7 @@ impl oio::BlockWrite for WebhdfsWriter {
 
         // rename concat file to path
         let resp = self
-            .backend
+            .core
             .webhdfs_rename_object(&first_block_id, &self.path)
             .await?;
 
@@ -132,7 +132,7 @@ impl oio::BlockWrite for WebhdfsWriter {
 
     async fn abort_block(&self, block_ids: Vec<Uuid>) -> Result<()> {
         for block_id in block_ids {
-            let resp = self.backend.webhdfs_delete(&block_id.to_string()).await?;
+            let resp = self.core.webhdfs_delete(&block_id.to_string()).await?;
             match resp.status() {
                 StatusCode::OK => {}
                 _ => return Err(parse_error(resp)),
@@ -144,7 +144,7 @@ impl oio::BlockWrite for WebhdfsWriter {
 
 impl oio::AppendWrite for WebhdfsWriter {
     async fn offset(&self) -> Result<u64> {
-        let resp = self.backend.webhdfs_get_file_status(&self.path).await?;
+        let resp = self.core.webhdfs_get_file_status(&self.path).await?;
         let status = resp.status();
         match status {
             StatusCode::OK => {
@@ -157,11 +157,11 @@ impl oio::AppendWrite for WebhdfsWriter {
             }
             StatusCode::NOT_FOUND => {
                 let req = self
-                    .backend
+                    .core
                     .webhdfs_create_object_request(&self.path, None, &self.op, Buffer::new())
                     .await?;
 
-                let resp = self.backend.client.send(req).await?;
+                let resp = self.core.client.send(req).await?;
                 let status = resp.status();
 
                 match status {
@@ -174,9 +174,9 @@ impl oio::AppendWrite for WebhdfsWriter {
     }
 
     async fn append(&self, _offset: u64, size: u64, body: Buffer) -> Result<Metadata> {
-        let location = self.backend.webhdfs_init_append_request(&self.path).await?;
-        let req = self.backend.webhdfs_append_request(&location, size, body)?;
-        let resp = self.backend.client.send(req).await?;
+        let location = self.core.webhdfs_init_append_request(&self.path).await?;
+        let req = self.core.webhdfs_append_request(&location, size, body)?;
+        let resp = self.core.client.send(req).await?;
 
         let status = resp.status();
         match status {
